refactor(vis,command-line-args): break circular type imports
Fixes three type-only file cycles identified by madge.

- command-line-args: resolve-args.ts imports from ./types, not ./index
- vis: extract ai types to ai-types.ts; ai-cache imports from there
- vis: move CodeownersConfig from codeowners.ts to workspace.ts
